Progress on Compilation
progress on compiling described in Compilation/11_16_2020 it's now trying to be done throught a Ubuntu vm compilation for android uzing buildozer, How to code trying to get done it's first episode, downloader stopped working. New idea where added, there is a description of how to use vim which will be a very useful coding utility in the near future. Research for Red-White option for mtg where also done it's in the idea factory