ideas, compilation
ideas: cerise au cognac pour notre trip au bc, an index of function is an idea where we upload simple searchable title in the form of questions like how to activate textwrap in notepad++ and then pour all those video in a channel then compilatory video with multiple chunks; downloadstream: we save our youtube search and all the interesting links in there to download the batch and store it in a media pool for offline use; I described a keypress problem with vlcM where it activate when pressing b in a text document for example it pauses the video I linked some video to download to help us with understanding how to patch that problem. I added the work we're doing right now with compiling the weed-app it'd be a good idea to update the weed-app repo with that too. If you ever read that to whomever spend a great day and enjoy your life :3 you're amzing <3